+++ pkgs/google_generative_ai/lib/src/model.dart
@@ -21,7 +21,7 @@ import 'client.dart';
 import 'content.dart';
 import 'function_calling.dart';
 
-const _apiVersion = 'v1';
+const _apiVersion = 'v1beta';
 Uri _googleAIBaseUri(RequestOptions? options) => Uri.https(
     'generativelanguage.googleapis.com', options?.apiVersion ?? _apiVersion);
 
@@ -40,8 +40,7 @@ enum Task {
 final class RequestOptions {
   /// The API version used to make requests.
   ///
-  /// By default the version is `v1`. This may be specified as `v1beta` to use
-  /// beta features.
+  /// By default the version is `v1beta`.
   final String? apiVersion;
   const RequestOptions({this.apiVersion});
 }
@@ -60,6 +59,7 @@ final class GenerativeModel {
   final ApiClient _client;
   final Uri _baseUri;
   final Content? _systemInstruction;
+  final ToolConfig? _toolConfig;
 
   /// Create a [GenerativeModel] backed by the generative model named [model].
   ///
@@ -85,12 +85,12 @@ final class GenerativeModel {
   /// request.
   ///
   /// Functions that the model may call while generating content can be passed
-  /// in [tools]. When using tools [requestOptions] must be passed to
-  /// override the `apiVersion` to `v1beta`.
+  /// in [tools]. Tool usage by the model can be configured with [toolConfig].
+  /// Tools and tool configuration can be overridden for individual requests
+  /// with arguments to [generateContent] or [generateContentStream].
   ///
   /// A [Content.system] can be passed to [systemInstruction] to give
-  /// high priority instructions to the model. When using system instructions
-  /// [requestOptions] must be passed to override the `apiVersion` to `v1beta`.
+  /// high priority instructions to the model.
   factory GenerativeModel({
     required String model,
     required String apiKey,
@@ -100,6 +100,7 @@ final class GenerativeModel {
     http.Client? httpClient,
     RequestOptions? requestOptions,
     Content? systemInstruction,
+    ToolConfig? toolConfig,
   }) =>
       GenerativeModel._withClient(
         client: HttpApiClient(apiKey: apiKey, httpClient: httpClient),
@@ -109,6 +110,7 @@ final class GenerativeModel {
         baseUri: _googleAIBaseUri(requestOptions),
         tools: tools,
         systemInstruction: systemInstruction,
+        toolConfig: toolConfig,
       );
 
   GenerativeModel._withClient({
@@ -119,12 +121,14 @@ final class GenerativeModel {
     required Uri baseUri,
     required List<Tool>? tools,
     required Content? systemInstruction,
+    required ToolConfig? toolConfig,
   })  : _model = _normalizeModelName(model),
         _baseUri = baseUri,
         _safetySettings = safetySettings,
         _generationConfig = generationConfig,
         _tools = tools,
         _systemInstruction = systemInstruction,
+        _toolConfig = toolConfig,
         _client = client;
 
   /// Returns the model code for a user friendly model name.
@@ -146,24 +150,35 @@ final class GenerativeModel {
   /// Sends a "generateContent" API request for the configured model,
   /// and waits for the response.
   ///
+  /// The [safetySettings], [generationConfig], [tools], and [toolConfig],
+  /// override the arguments of the same name passed to the
+  /// [GenerativeModel.new] constructor. Each argument, when non-null,
+  /// overrides the model level configuration in its entirety.
+  ///
   /// Example:
   /// ```dart
   /// final response = await model.generateContent([Content.text(prompt)]);
   /// print(response.text);
   /// ```
-  Future<GenerateContentResponse> generateContent(Iterable<Content> prompt,
-      {List<SafetySetting>? safetySettings,
-      GenerationConfig? generationConfig}) async {
+  Future<GenerateContentResponse> generateContent(
+    Iterable<Content> prompt, {
+    List<SafetySetting>? safetySettings,
+    GenerationConfig? generationConfig,
+    List<Tool>? tools,
+    ToolConfig? toolConfig,
+  }) async {
     safetySettings ??= _safetySettings;
     generationConfig ??= _generationConfig;
+    tools ??= _tools;
+    toolConfig ??= _toolConfig;
     final parameters = {
       'contents': prompt.map((p) => p.toJson()).toList(),
       if (safetySettings.isNotEmpty)
         'safetySettings': safetySettings.map((s) => s.toJson()).toList(),
-      if (generationConfig case final config?)
-        'generationConfig': config.toJson(),
-      if (_tools case final tools?)
-        'tools': tools.map((t) => t.toJson()).toList(),
+      if (generationConfig != null)
+        'generationConfig': generationConfig.toJson(),
+      if (tools != null) 'tools': tools.map((t) => t.toJson()).toList(),
+      if (toolConfig != null) 'toolConfig': toolConfig.toJson(),
       if (_systemInstruction case final systemInstruction?)
         'systemInstruction': systemInstruction.toJson(),
     };
@@ -177,6 +192,11 @@ final class GenerativeModel {
   /// Sends a "streamGenerateContent" API request for the configured model,
   /// and waits for the response.
   ///
+  /// The [safetySettings], [generationConfig], [tools], and [toolConfig],
+  /// override the arguments of the same name passed to the
+  /// [GenerativeModel.new] constructor. Each argument, when non-null,
+  /// overrides the model level configuration in its entirety.
+  ///
   /// Example:
   /// ```dart
   /// final responses = await model.generateContent([Content.text(prompt)]);
@@ -185,19 +205,24 @@ final class GenerativeModel {
   /// }
   /// ```
   Stream<GenerateContentResponse> generateContentStream(
-      Iterable<Content> prompt,
-      {List<SafetySetting>? safetySettings,
-      GenerationConfig? generationConfig}) {
+    Iterable<Content> prompt, {
+    List<SafetySetting>? safetySettings,
+    GenerationConfig? generationConfig,
+    List<Tool>? tools,
+    ToolConfig? toolConfig,
+  }) {
     safetySettings ??= _safetySettings;
     generationConfig ??= _generationConfig;
+    tools ??= _tools;
+    toolConfig ??= _toolConfig;
     final parameters = <String, Object?>{
       'contents': prompt.map((p) => p.toJson()).toList(),
       if (safetySettings.isNotEmpty)
         'safetySettings': safetySettings.map((s) => s.toJson()).toList(),
-      if (generationConfig case final config?)
-        'generationConfig': config.toJson(),
-      if (_tools case final tools?)
-        'tools': tools.map((t) => t.toJson()).toList(),
+      if (generationConfig != null)
+        'generationConfig': generationConfig.toJson(),
+      if (tools != null) 'tools': tools.map((t) => t.toJson()).toList(),
+      if (toolConfig != null) 'toolConfig': toolConfig.toJson(),
       if (_systemInstruction case final systemInstruction?)
         'systemInstruction': systemInstruction.toJson(),
     };
@@ -290,8 +315,9 @@ GenerativeModel createModelWithClient({
   List<SafetySetting> safetySettings = const [],
   GenerationConfig? generationConfig,
   RequestOptions? requestOptions,
-  List<Tool>? tools,
   Content? systemInstruction,
+  List<Tool>? tools,
+  ToolConfig? toolConfig,
 }) =>
     GenerativeModel._withClient(
       client: client,
@@ -299,8 +325,9 @@ GenerativeModel createModelWithClient({
       safetySettings: safetySettings,
       generationConfig: generationConfig,
       baseUri: _googleAIBaseUri(requestOptions),
-      tools: tools,
       systemInstruction: systemInstruction,
+      tools: tools,
+      toolConfig: toolConfig,
     );
 
 /// Creates a model with an overridden base URL to communicate with a different
@@ -313,6 +340,7 @@ GenerativeModel createModelWithBaseUri({
   required String model,
   required String apiKey,
   required Uri baseUri,
+  FutureOr<Map<String, String>> Function()? requestHeaders,
   List<SafetySetting> safetySettings = const [],
   GenerationConfig? generationConfig,
   List<Tool>? tools,
@@ -320,12 +348,11 @@ GenerativeModel createModelWithBaseUri({
   ToolConfig? toolConfig,
 }) =>
     GenerativeModel._withClient(
-      client: HttpApiClient(apiKey: apiKey),
+      client: HttpApiClient(apiKey: apiKey, requestHeaders: requestHeaders),
       model: model,
       safetySettings: safetySettings,
       generationConfig: generationConfig,
       baseUri: baseUri,
-      tools: null,
       systemInstruction: systemInstruction,
       tools: tools,
       toolConfig: toolConfig,